v0.0.3.5 update

+++ PINGPONG_script/upload.py
@@ -1,6 +1,11 @@
 import os
 import shutil
 import time
+import sys
+import traceback
+from pathlib import Path
+sys.path.append("..")
+from main import print_good, print_normal
 def Upload(ip, file_dir, to_dir, printf, APP_SEND, conn, sendobj):
     try:
         if APP_SEND:
@@ -9,7 +14,7 @@ def Upload(ip, file_dir, to_dir, printf, APP_SEND, conn, sendobj):
             IS_C = True
         if IS_C:
             if printf:
-                print("PINGPONG>[*]Sending dir......")
+                print_normal("PINGPONG>[*]Sending dir......")
             conn.send(bytes(to_dir, "utf8"))
             path = "temp"
             if os.path.exists(path):
@@ -17,12 +22,20 @@ def Upload(ip, file_dir, to_dir, printf, APP_SEND, conn, sendobj):
             if not os.path.exists(path):
                 os.makedirs(path)
             if printf:
-                print("PINGPONG>[*]Copy to file to " + path + "......")
+                print_normal("PINGPONG>[*]Copy to file to " + path + "......")
             #判断是否为文件/文件夹
             if not os.path.isfile(file_dir):
                 try:
                     file_names = os.listdir(file_dir)
+                    dir_list = Path(file_dir)
+                    dirs = [e for e in dir_list.iterdir() if e.is_dir()]
+                    for dir in dirs:
+                        na = dir.resolve()
+                        start = str(na).rindex('/')
+                        name = str(na)[start+1:]
+                        file_names.remove(name)
                 except:
+                    print(traceback.print_exc())
                     ch = input("PINGPONG>[-]The location is NOT valuable, again?[y/n]>")
                     if ch == "y" or ch == "yes" or ch == "YES" or ch == "Y":
                         file_dir = input("PINGPONG>[*]Please input the location of the file in your host>")
@@ -31,40 +44,42 @@ def Upload(ip, file_dir, to_dir, printf, APP_SEND, conn, sendobj):
                     else:
                         return True
                 for fi in os.listdir(file_dir):
-                    full_file_name = os.path.join(file_dir, fi)
-                    if os.path.isfile(full_file_name):
-                        shutil.copy(full_file_name, path)
+                    if not os.path.isdir(fi):
+                        full_file_name = os.path.join(file_dir, fi)
+                        if os.path.isfile(full_file_name):
+                            shutil.copy(full_file_name, path)
                 for file in file_names:
-                    old_n = os.path.join(path, file)
-                    new_name = file + ".txt"
-                    os.rename(old_n, new_name)
-                    with open(new_name, "rb") as f:
-                        se_data = f.read()
-                        conn.send(bytes(new_name, "utf8"))
-                        name_data = conn.recv(1024)
-                        while True:
-                            time.sleep(1)
-                            if name_data:
-                                break
-                        len_data = len(se_data)
-                        conn.send(bytes(str(len_data), "utf8"))
-                        len_recv = conn.recv(1024)
-                        while True:
-                            time.sleep(1)
-                            if len_recv:
-                                break
-                        conn.sendall(se_data)
-                        upload_data = conn.recv(1024)
-                        while True:
-                            time.sleep(1)
-                            if upload_data:
-                                print(ip + ">[*]File Upload Succeed: " + old_n + " >>> " + to_dir + "/" + file)
-                                break
-                        f.close()
-                        try:
-                            os.remove(new_name)
-                        except:
-                            pass
+                    if not os.path.isdir(file):
+                        old_n = os.path.join(path, file)
+                        new_name = file + ".txt"
+                        os.rename(old_n, new_name)
+                        with open(new_name, "rb") as f:
+                            se_data = f.read()
+                            conn.send(bytes(new_name, "utf8"))
+                            name_data = conn.recv(1024)
+                            while True:
+                                time.sleep(1)
+                                if name_data:
+                                    break
+                            len_data = len(se_data)
+                            conn.send(bytes(str(len_data), "utf8"))
+                            len_recv = conn.recv(1024)
+                            while True:
+                                time.sleep(1)
+                                if len_recv:
+                                    break
+                            conn.sendall(se_data)
+                            upload_data = conn.recv(1024)
+                            while True:
+                                time.sleep(1)
+                                if upload_data:
+                                    print_normal(ip + ">[*]File Upload Succeed: " + old_n + " >>> " + to_dir + "/" + file)
+                                    break
+                            f.close()
+                            try:
+                                os.remove(new_name)
+                            except:
+                                pass
             else:
                 file = file_dir[file_dir.rindex('/') + 1:len(file_dir)]
                 try:
@@ -81,7 +96,7 @@ def Upload(ip, file_dir, to_dir, printf, APP_SEND, conn, sendobj):
                 file_name = file_dir.split(".txt")[0]
                 file_names =  file_name[file_name.rindex('/') + 1:len(file_name)]
                 if printf:
-                    print("PINGPONG>[*]Sending file......")
+                    print_normal("PINGPONG>[*]Sending file......")
                 f = open(file_names + ".txt", "rb")    
                 se_data = f.read()
                 conn.send(bytes(new_name, "utf8"))
@@ -102,20 +117,20 @@ def Upload(ip, file_dir, to_dir, printf, APP_SEND, conn, sendobj):
                 while True:
                     time.sleep(1)
                     if upload_data:
-                        print(ip + ">[*]File Upload Succeed: " + old_n + " >>> " + to_dir + "/" + file)
+                        print_normal(ip + ">[*]File Upload Succeed: " + old_n + " >>> " + to_dir + "/" + file)
                         break
                 f.close()
                 try:
                     os.remove(new_name)
                 except:
                     pass
-        # for check_d in file_names:
-        #     if os.path.isdir(check_d):
-        #         Upload(os.path.join(file_dir, check_d), os.path.join(to_dir, check_d), False, False)
+        for check_d in file_names:
+            if os.path.isdir(check_d):
+                Upload(os.path.join(file_dir, check_d), os.path.join(to_dir, check_d), False, False)
         conn.send(bytes("END", 'utf8'))
-        print("PINGPONG>[+]FILE UPLOAD DONE")
+        print_good("PINGPONG>[+]FILE UPLOAD DONE")
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         restart = input("PINGPONG>[-]Something went WRONG, restart?[y/n]")
         if restart == "y" or restart == "yes" or restart == "YES" or restart == "Y":
             file_dir = input("PINGPONG>[*]Please input the location of the file in your host>")
